last commit

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,17 @@ use DB;
 
 class PostsController extends Controller
 {
+
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -79,6 +90,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Access');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -112,6 +128,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Access');
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success','Post Has Been Deleted !');
